Abhya-main/requirements.txt
absl-py 
appnope 
argon2-cffi 
astor==0.8.1
async-generator==1.10
attrs 
backcall 
bleach 
certifi==2020.12.5
cffi 
chardet==4.0.0
click==7.1.2
contrib==0.3.0
coverage 
cycler==0.10.0
Cython 
decorator 
defusedxml 
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio 
h5py 
idna==2.10
imageio==2.9.0
importlib-metadata 
itsdangerous==1.1.0
jedi 
Jinja2 
jsonschema 
Keras-Applications 
Keras-Preprocessing 
kiwisolver==1.3.1
Markdown 
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient 
nbconvert 
nbformat 
nest-asyncio 
networkx==2.5
notebook 
numpy 
opencv-python==4.5.1.48
opt-einsum==3.1.0
packaging 
pandocfilters 
parso==0.7.0
pexpect 
pickleshare 
Pillow==8.1.2
prometheus-client 
prompt-toolkit 
protobuf==3.14.0
ptyprocess 
pycparser 
Pygments 
pyparsing 
pyrsistent 
python-dateutil 
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole 
QtPy==1.9.0
requests==2.25.1
scikit-image
scipy 
Send2Trash 
six 
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.9.2
testpath 
tifffile
tornado 
traitlets 
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth 
webencodings==0.5.1
Werkzeug 
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp 
